requirements.txt
Django==4.0
django-cors-headers==3.10.1
mysqlclient==2.1.0
PyJWT==2.3.0
django-extensions==3.1.5
ipython==7.30.1
requests==2.26.0
